third_party/WebKit/LayoutTests/dom/domparsing/xmlserializer-prebound-xml-prefix.html

<!DOCTYPE html>
<title>XMLSerializer: Pre-bound namespace prefixes</title>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<div id=log></div>
<script>
test(function() {
    var s = '<foo xml:space="default"/>';
    var docFrag = new DOMParser().parseFromString(s, 'text/xml');
    var root = docFrag.removeChild(docFrag.firstChild);
    var result = new XMLSerializer().serializeToString(root);
    assert_equals(result, '<foo xml:space="default"/>');
}, 'XMLSerializer does not emit a namespace declaration for the \'xml\' prefix.');
</script>
